feat: Refactor data fetching and implement testing infrastructure
This commit refactors the data fetching mechanism using a generic `networkBoundResource` and introduces a comprehensive testing setup.

- A `core:testing` module has been created to house testing utilities, including `FakeAppDataBase`, `FakeForecastDailyDao`, and Koin test modules.
- The `ForecastDailyRepositoryImpl` is updated to use the new `networkBoundResource` for a cleaner single source of truth implementation. This resource now handles fetching data, saving it to the local database, and emitting results from the database as a flow.
- Outdated weather data is now cleared using a new `DateUtils` utility.
- The `ForecastDailyRemoteDataSource` is now an interface with a Ktor implementation, and its DI bindings have been updated.
- Added the `INTERNET` permission to the `AndroidManifest.xml`.
- Implemented comprehensive unit tests for `ForecastDailyRepository` to verify caching, fetching, and error handling logic, using Turbine and Truth.
- DTOs (`DailyDto`, `FeelsLikeDto`) have been updated to use `Float` instead of `Double` for consistency.
- Test dependencies have been organized into bundles (`local-tests`, `instrumented-tests`) in `libs.versions.toml` for better management.

+++ feature/forecast_daily/data/src/main/java/com/panabuntu/weathertracker/feature/forecast_daily/data/remote_data_source/dto/DailyDto.kt
@@ -9,7 +9,7 @@ data class DailyDto(
     @SerialName("clouds")
     val clouds: Int,
     @SerialName("dew_point")
-    val dewPoint: Double,
+    val dewPoint: Float,
     @SerialName("dt")
     val dt: Long,
     @SerialName("feels_like")
@@ -17,17 +17,17 @@ data class DailyDto(
     @SerialName("humidity")
     val humidity: Int,
     @SerialName("moon_phase")
-    val moonPhase: Double,
+    val moonPhase: Float,
     @SerialName("moonrise")
     val moonrise: Int,
     @SerialName("moonset")
     val moonset: Int,
     @SerialName("pop")
-    val pop: Double,
+    val pop: Float,
     @SerialName("pressure")
     val pressure: Int,
     @SerialName("rain")
-    val rain: Double,
+    val rain: Float,
     @SerialName("summary")
     val summary: String,
     @SerialName("sunrise")
@@ -37,13 +37,13 @@ data class DailyDto(
     @SerialName("temp")
     val temp: TempDto,
     @SerialName("uvi")
-    val uvi: Double,
+    val uvi: Float,
     @SerialName("weather")
     val weather: List<WeatherDto>,
     @SerialName("wind_deg")
     val windDeg: Int,
     @SerialName("wind_gust")
-    val windGust: Double,
+    val windGust: Float,
     @SerialName("wind_speed")
-    val windSpeed: Double
+    val windSpeed: Float
 )
