+++ feature/forecast_daily/data/src/main/java/com/panabuntu/weathertracker/feature/forecast_daily/data/remote_data_source/dto/FeelsLikeDto.kt
@@ -7,11 +7,11 @@ import kotlinx.serialization.Serializable
 @Serializable
 data class FeelsLikeDto(
     @SerialName("day")
-    val day: Double,
+    val day: Float,
     @SerialName("eve")
-    val eve: Double,
+    val eve: Float,
     @SerialName("morn")
-    val morn: Double,
+    val morn: Float,
     @SerialName("night")
-    val night: Double
+    val night: Float
 )
